nlg-backend/requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.2.5
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==2.9.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
Markdown==3.4.1
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
